+++ view/frontend/web/js/view/payment.js
@@ -30,11 +30,21 @@ define(['jquery', 'ko', 'Magento_Checkout/js/view/payment/default'], function (
       customerCards:
         window.checkoutConfig.payment.hiPayFullservice.customerCards,
       createOneclick: false,
+      maxSavedCard: window.checkoutConfig.payment.hiPayFullservice.maxSavedCard,
       creditCardType: '',
+      creditCardOwner: '',
+      creditCardNumber: '',
+      creditCardExpMonth: '',
+      creditCardExpYear: '',
+      multiUse: '',
       fingerprint: '',
       defaultEci: window.checkoutConfig.payment.hiPayFullservice.defaultEci,
       recurringEci: window.checkoutConfig.payment.hiPayFullservice.recurringEci,
       eci: window.checkoutConfig.payment.hiPayFullservice.defaultEci,
+      availableBrands:
+          window.checkoutConfig.payment.hipay_hosted_fields !== undefined
+              ? window.checkoutConfig.payment.hipay_hosted_fields.availableTypes
+              : '',
       showForm: true
     },
     getAfterPlaceOrderUrl: function () {
@@ -46,7 +56,12 @@ define(['jquery', 'ko', 'Magento_Checkout/js/view/payment/default'], function (
         'selectedCard',
         'createOneclick',
         'creditCardType',
+        'creditCardOwner',
+        'creditCardNumber',
+        'creditCardExpMonth',
+        'creditCardExpYear',
         'creditCardToken',
+        'multiUse',
         'eci',
         'fingerprint'
       ]);
@@ -102,6 +117,11 @@ define(['jquery', 'ko', 'Magento_Checkout/js/view/payment/default'], function (
     getCustomerCards: function () {
       return this.customerCards;
     },
+
+    getCustomerSavedCardsCount: function () {
+      return this.maxSavedCard[this.getCode()];
+    },
+
     getCustomerCardByToken: function (token) {
       for (var i = 0; i < this.customerCards.length; i++) {
         if (this.customerCards[i].ccToken == token) {
@@ -114,9 +134,39 @@ define(['jquery', 'ko', 'Magento_Checkout/js/view/payment/default'], function (
       return this.allowOneclick[this.getCode()];
     },
 
-    customerHasCard: function () {
-      return this.getCustomerCards().length > 0;
+    customerHasCard: function() {
+      let customerCards = this.getCustomerCards();
+      let availableBrands = this.getAvailableBrands(this.availableBrands);
+
+      return customerCards.length > 0 &&
+          customerCards.some(card => availableBrands.includes(card.brand));
     },
+
+    getAvailableBrands: function (brand)  {
+      let result = new Set();
+
+      Object.entries(brand).forEach(([key, value]) => {
+        switch(key) {
+          case 'VI':
+            result.add('visa');
+            result.add('cb');
+            break;
+          case 'MC':
+            result.add('mastercard');
+            break;
+          case 'AE':
+            result.add('american-express');
+            break;
+          case 'MI':
+            result.add('maestro');
+            result.add('bcmc');
+            break;
+        }
+      });
+
+      return Array.from(result);
+    },
+
     getData: function () {
       var fingerprint = $('#ioBB').val();
 
@@ -129,6 +179,11 @@ define(['jquery', 'ko', 'Magento_Checkout/js/view/payment/default'], function (
         additional_data: {
           create_oneclick: this.createOneclick(),
           card_token: this.creditCardToken(),
+          card_owner: this.creditCardOwner(),
+          card_pan: this.creditCardNumber(),
+          card_expiry_month: this.creditCardExpMonth(),
+          card_expiry_year: this.creditCardExpYear(),
+          card_multi_use: this.multiUse(),
           eci: this.eci(),
           cc_type: this.creditCardType(),
           fingerprint: fingerprint