+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
+## 1.27.0
+
+- **Add** : Update OneClick payment
+- **Fix** : Cancel order when payment abandonment instead of updated to pending state
+
 ## 1.26.0
 
+- **Fix** : Lock order table row while executing notification
 - **Add** : Dynamically get Alma MAX/MIN from merchant account
 - **Fix** : Lock order table row while executing notification
 - **Fix** : Allow retrieving customer X-Forwarded-For IP addresse